moving the time test boilerplate stuff to the old util
